requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
loguru==0.6.0
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
psycopg2==2.9.5
pycparser==2.21
pyhcl==0.4.4
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
shortuuid==1.0.11
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
validators==0.20.0
